6.秒杀接口的优化

+++ src/main/java/com/imooc/miaosha/dao/OrderDao.java
@@ -17,7 +17,7 @@ public interface OrderDao {
             ",#{status},#{createDate})")
     //通过SelectKey返回对应的Id,keyProperty代表的是domain对象中对应column的属性
     @SelectKey(keyColumn = "id",keyProperty = "id",resultType = Long.class,before = false,statement = "select last_insert_id()")
-    public long insert(OrderInfo orderInfo);
+    public int insert(OrderInfo orderInfo);
     @Insert("insert into miaosha_order(user_id,goods_id,order_id) " +
             "values(#{userId},#{goodsId},#{orderId})")
     public int insertMiaoshaOrder(MiaoshaOrder miaoshaOrder);
